began refactoring for database seeding
