+++ client/src/components/PledgeInfo.jsx
@@ -20,7 +20,7 @@ class PledgeInfo extends Component {
     if (this.props.id) {
       this.getData(this.props.id);
     } else {
-      this.getData(6);
+      this.getData(8);
     }
   }
