+++ client/src/components/PledgeSupport.jsx
@@ -17,7 +17,7 @@ class PledgeSupport extends React.Component {
     if (this.props.id) {
       this.getData(this.props.id);
     } else {
-      this.getData(6);
+      this.getData(8);
     }
   }
